LayoutTests/fast/text-extraction/debug-text-extraction-scrolling-expected.txt

root,scrollPosition=(0,0),contentSize=[…]
    scrollable,scrollPosition=(5000,0),contentSize=[…]
        link,url='https://webkit.org/'
    scrollable,scrollPosition=(10,0),contentSize=[…],'2. Hello world'
